Modified bayesian.jpi file for triggering repo scan for CVEs
A new injection is added (gitURL) for triggering repo scan in bayesian jenkins plugin.
This commit is to affect that change and the .hpi file (renamed to .jpi) file was added
to this repo.

The .hpi file was built from fabric8-analytics/fabric8-analytics-jenkins-plugin#14 tag:v1.0.1
Issue: openshiftio/openshift.io#3925